+++ core/src/test/scala/spinal/core/ClockDomain.scala
@@ -3,7 +3,7 @@ package spinal.core
 import scala.util.Random
 
 import spinal.core.sim._
-import spinal.lib.{BufferCC, Counter, CounterFreeRun, NoData}
+import spinal.lib.{BufferCC, Counter, CounterFreeRun, NoData, ResetCtrl}
 import spinal.tester.{SpinalSimFunSuite, SpinalTesterCocotbBase}
 
 object ClockDomainConfigTester {
@@ -480,4 +480,41 @@ class InternalClockTesterCocotbBoot extends SpinalTesterCocotbBase {
   override def getName: String = "InternalClockTester"
   override def pythonTestLocation: String = "tester/src/test/python/spinal/InternalClockTester"
   override def createToplevel: Component = new InternalClockTester
+}
+
+object PllAAssertSDeassertTester{
+  class PLL extends BlackBox{
+    val io = new Bundle{
+      val clk_in = in Bool()
+      val clk_out = out Bool()
+    }
+
+    noIoPrefix()
+  }
+
+  class PllAAssertSDeassertTester extends Component{
+    val aReset     = in Bool()
+    val clk_100Mhz = in Bool()
+
+    val pllBB = new PLL
+    pllBB.io.clk_in := clk_100Mhz
+
+    val coreClockDomain = ClockDomain.internal("core")
+    coreClockDomain.clock := pllBB.io.clk_out
+    coreClockDomain.reset := ResetCtrl.asyncAssertSyncDeassert(
+      input       = aReset,
+      clockDomain = coreClockDomain
+    )
+
+    val counter = out(coreClockDomain(CounterFreeRun(10).value)).setName("counter")
+  }
+}
+
+class PllAAssertSDeassertTesterBoot extends SpinalTesterCocotbBase {
+  override def getName: String = "PllAAssertSDeassertTester"
+  override def pythonTestLocation: String = "tester/src/test/python/spinal/PllAAssertSDeassertTester"
+  override def createToplevel: Component = new PllAAssertSDeassertTester.PllAAssertSDeassertTester
+
+  override def backendConfig(config: SpinalConfig): SpinalConfig = config
+    .copy(defaultConfigForClockDomains = config.defaultConfigForClockDomains.copy(resetKind = SYNC))
 }
